Move `RegexGlobal` to Ocelot.Infrastructure namespace
