+++ src/Ocelot/Authorization/ClaimsAuthorizer.cs
@@ -1,4 +1,5 @@
 ﻿using Ocelot.DownstreamRouteFinder.UrlMatcher;
+using Ocelot.Infrastructure;
 using Ocelot.Infrastructure.Claims.Parser;
 using Ocelot.Responses;
 using System.Security.Claims;
