+++ src/Ocelot/Configuration/Creator/UpstreamHeaderTemplatePatternCreator.cs
@@ -1,4 +1,5 @@
 using Ocelot.Configuration.File;
+using Ocelot.Infrastructure;
 using Ocelot.Values;
 
 namespace Ocelot.Configuration.Creator;
