+++ src/Ocelot/DependencyInjection/ConfigurationBuilderExtensions.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Configuration.Memory;
 using Newtonsoft.Json;
 using Ocelot.Configuration.File;
+using Ocelot.Infrastructure;
 
 namespace Ocelot.DependencyInjection
 {
