+++ src/Ocelot/Values/UpstreamPathTemplate.cs
@@ -1,3 +1,5 @@
+using Ocelot.Infrastructure;
+
 namespace Ocelot.Values
 {
     public class UpstreamPathTemplate
